lab_3/requirements1.txt
annotated-types==0.6.0
autocommand==2.2.2
cheroot==10.0.0
CherryPy==18.9.0
contourpy==1.2.0
cycler==0.12.1
DataSpyre==0.2.8
DateTime==5.4
fonttools==4.47.0
inflect==7.0.0
jaraco.collections==5.0.0
jaraco.context==4.3.0
jaraco.functools==4.0.0
jaraco.text==3.12.0
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
more-itertools==10.1.0
numpy==1.26.2
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
portend==3.2.0
pydantic==2.5.2
pydantic_core==2.14.5
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
tempora==5.5.0
typing_extensions==4.9.0
tzdata==2023.3
zc.lockfile==3.0.post1
zope.interface==6.1
